requirements.txt
flask
gunicorn
gTTS
deep-translator
edge-tts
google-generativeai
